Change barcode sort order to match cellranger
This is a very small change to help make more outputs of the count
WDL bit-identical to the outputs from the cellranger tarball. The
pairs of bases always appear to show up in the same order when joining,
and though it really shouldn't make a difference, this appears to
affect the output if we use a different order.