Use jq slurpfile for possibly large variables
Previously, we were using command substitution and redirection
to insert the contents of json files in the _args jsons. But
when the contents of that file are really large, you get
problems like "argument list too large". This change replaces
that with the slurpfile option from jq, which reads a json file
and doesn't make you send its contents through the command line.