Improve SmartSeq2 portability
Four changes from running it with dxWDL:

1. Remove unused input from json

Other environments won't know that this (very large) input file isn't
used by the workflow unless they parse the WDL prior to attempting any
file localization.

2. Switch from File to String for string inputs

3. Don't try to preserve ownership when untarring

When you untar an archive as root, it tried to keep uid and gid
ownership the same. But that could fail in some environments, and it
doesn't really help us, so this change allows the files to be extracted
as the docker user.

4. Set memory requirement to the actual google instance

These tasks request 4 cores and 5GB of memory. On GCP, this should get
you an n1-standard-4, which actually has 15GB of memory. On AWS, you
might get a c4.xlarge, which has 7.5GB of memory.

Some of the tasks appear to actually require between 7.5 and 15GB of
memory, so they succeed on GCP but fail on AWS even though both
platforms provide an instance that meets the declared requirements.