Add "preemptible" input to Optimus workflow
Default value of the "preemptible" input is 3.
This input is passed to ALL tasks called by
the Optimus workflow.

Default "preemptible" input in all tasks was also
set to 3, changed from 0.

The "preemptible" input was also added to
ZarrUtils.wdl, where it was previously
described in the parameters metadata section
but not actually implemented.