Add SNAP-ATAC Pipeline in WDL form (#211)
This commit adds the SNAP-ATAC pipeline in WDL form along with a workflow for generating references for this pipeline. The pipeline is very close to the original SNAP-ATAC pipeline but contains tests, provides output as individual text files in addition to a snap file and also provides a BAM where the cell barcodes are in tags and not the read names. The pipeline is accompanied by the README file.